Create one_x_multi_y_plot.py
This script first prompts the user to enter the column numbers separated by commas using the input function and uses the map and split functions to store the values in a list called column_numbers. Then, it takes the file_list_path as an argument using argparse library. Then, it opens the file containing the list of file paths using the open function in read mode. Then it reads the list of file paths using the readlines method and removes the newline characters from each path using a list comprehension. Then, it iterates over the list of file paths, checking if each path is a valid file path using the os.path.isfile function.