Create plot_multicsv_for_notebook_jupyter.py
In this script, it first defines the variable file_list_path and column_numbers directly in the script and uses them to open the file, read the contents of each file, writes the contents of specific columns as many as desired and plots the multiple data with the first column as the x-axis and the rest of the each columns as the y-axis.
It also plots the data using matplotlib library, where x_values will be used for x-axis and y_values for y-axis. It also gives title, and labels for x-axis and y-axis. Finally, it shows the plot using plt.show()